backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
Pillow==10.1.0
PyYAML==6.0.1
typer==0.9.0
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3